wrote tests for the bug (piece equality with null)
